requirements.txt
# Core Dependencies - Required for basic functionality
click>=8.0.0
sqlalchemy>=2.0.0
pyyaml>=6.0.0
pillow>=9.0.0

# Optional Dependencies - Install as needed
# For API server functionality:
# flask>=2.3.0
# flask-cors>=4.0.0

# For Excel export functionality:
# pandas>=1.5.0
# openpyxl>=3.1.0

# For development and testing:
pytest>=7.0.0
pytest-cov>=4.0.0

# SQLite is included with Python - no additional database drivers needed
# psycopg2-binary>=2.9.0
# python-dotenv>=1.0.0
